requirements.txt
langgraph
langchain
langchain-openai
langchain-community
chromadb
beautifulsoup4
requests
python-dotenv
openai
click
